Completed File Upload Section

+++ controllers/admin.js
@@ -1,6 +1,8 @@
 const Product = require('../models/product');
 const { validationResult } = require('express-validator');
 
+const fileHelper = require('../util/file');
+
 exports.getAddProduct = (req, res, next) => {
 	res.render('admin/edit-product', {
 		pageTitle: 'Add Product',
@@ -14,9 +16,24 @@ exports.getAddProduct = (req, res, next) => {
 
 exports.postAddProduct = (req, res, next) => {
 	const title = req.body.title;
-	const imageUrl = req.body.imageUrl;
+	const image = req.file;
 	const price = req.body.price;
 	const description = req.body.description;
+	if (!image) {
+		return res.status(422).render('admin/edit-product', {
+			pageTitle: 'Add Product',
+			path: '/admin/add-product',
+			editing: false,
+			hasError: true,
+			product: {
+				title: title,
+				price: price,
+				description: description,
+			},
+			errorMessage: 'The file uploaded is not a image',
+			validationErrors: [],
+		});
+	}
 	const errors = validationResult(req);
 	if (!errors.isEmpty()) {
 		return res.status(422).render('admin/edit-product', {
@@ -26,7 +43,6 @@ exports.postAddProduct = (req, res, next) => {
 			hasError: true,
 			product: {
 				title: title,
-				imageUrl: imageUrl,
 				price: price,
 				description: description,
 			},
@@ -34,6 +50,7 @@ exports.postAddProduct = (req, res, next) => {
 			validationErrors: errors.array(),
 		});
 	}
+	const imageUrl = image.path;
 	const product = new Product({
 		title: title,
 		price: price,
@@ -81,7 +98,7 @@ exports.postEditProduct = (req, res, next) => {
 	const prodId = req.body.productId;
 	const updatedTitle = req.body.title;
 	const updatedPrice = req.body.price;
-	const updatedImageUrl = req.body.imageUrl;
+	const image = req.file;
 	const updatedDesc = req.body.description;
 	const errors = validationResult(req);
 	if (!errors.isEmpty()) {
@@ -92,7 +109,6 @@ exports.postEditProduct = (req, res, next) => {
 			hasError: true,
 			product: {
 				title: updatedTitle,
-				imageUrl: updatedImageUrl,
 				price: updatedPrice,
 				description: updatedDesc,
 				_id: prodId,
@@ -109,7 +125,10 @@ exports.postEditProduct = (req, res, next) => {
 			product.title = updatedTitle;
 			product.price = updatedPrice;
 			product.description = updatedDesc;
-			product.imageUrl = updatedImageUrl;
+			if (image) {
+				fileHelper.deleteFile(product.imageUrl);
+				product.imageUrl = image.path;
+			}
 			return product.save().then(result => {
 				console.log('UPDATED PRODUCT!');
 				res.redirect('/admin/products');
@@ -134,10 +153,17 @@ exports.getProducts = (req, res, next) => {
 
 exports.postDeleteProduct = (req, res, next) => {
 	const prodId = req.body.productId;
-	Product.deleteOne({ _id: prodId, userId: req.user._id })
+	Product.findById(prodId)
+		.then(product => {
+			if (!product) {
+				return next(new Error('No products found'));
+			}
+			fileHelper.deleteFile(product.imageUrl);
+			return Product.deleteOne({ _id: prodId, userId: req.user._id });
+		})
 		.then(() => {
 			console.log('DESTROYED PRODUCT');
 			res.redirect('/admin/products');
 		})
-		.catch(err => console.log(err));
+		.catch(err => next(err));
 };
